eval-dashboard/requirements.txt
fastapi
pymongo[srv]
python-dotenv
uvicorn
